+++ activerecord/lib/active_record/relation.rb
@@ -418,8 +418,7 @@ module ActiveRecord
         records.each { |record| record.update(attributes) }
       else
         if ActiveRecord::Base === id
-          id = id.id
-          ActiveSupport::Deprecation.warn(<<-MSG.squish)
+          raise ArgumentError, <<-MSG.squish
             You are passing an instance of ActiveRecord::Base to `update`.
             Please pass the id of the object by calling `.id`.
           MSG